:sparkles: you never know when it's gonna crash on you

+++ Labo_2020/IPointy.cs
@@ -8,6 +8,6 @@ namespace Labo_2020
 {
 	public interface IPointy
 	{
-		byte NbPoints { get; }
+		int NbPoints { get; }
 	}
 }
